debug

+++ hddm/models/hddm_rl.py
@@ -612,8 +612,8 @@ class HDDMrl(HDDM):
             wfpt_parents['window_start'] = -1.00
             wfpt_parents['window_size'] = -1.00
         else:
-            wfpt_parents['window_start'] = self.window_start.astype(float)
-            wfpt_parents['window_size'] = self.window_size.astype(float)
+            wfpt_parents['window_start'] = self.window_start
+            wfpt_parents['window_size'] = self.window_size
 
         # wfpt_parents["z"] = knodes["z_bottom"] if "z" in self.include else 0.5
